#227 moved dom tests to packages
