+++ packages/react-look/test/index-test.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react'
+
+import look from '../modules/index'
+
+describe('Enhancing with look', () => {
+
+  it('should work as a higher order function', () => {
+    const Comp = () => <div />
+    expect(look(Comp).displayName).to.eql('Comp')
+  })
+})
