+++ packages/react-look/test/mocha.opts
@@ -1,0 +1,3 @@
+--compilers js:babel-core/register
+--recursive
+--require test/setup.js
