+++ packages/react-look/test/setup.js
@@ -1,0 +1,2 @@
+var chai = require('chai')
+global.expect = chai.expect
