+++ packages/react-look/test/utils/camelToDashCase-test.js
@@ -1,5 +1,4 @@
 import camelToDashCase from '../../modules/utils/camelToDashCase'
-import { expect } from 'chai'
 
 describe('Converting camel case to dash case', () => {
