+++ packages/react-look/test/utils/getFontFormat-test.js
@@ -1,5 +1,4 @@
 import getFontFormat from '../../modules/utils/getFontFormat'
-import { expect } from 'chai'
 
 describe('Evaluating the font format', () => {
