--- test/core/dom/resolver-test.js
@@ -1,10 +1,0 @@
-import { expect } from 'chai'
-import React, { Component } from 'react'
-import StyleSheet from '../../../modules/api/dom/StyleSheet'
-import resolveStyles from '../../../modules/core/dom/resolver'
-import look from '../../../modules/core/enhancer'
-import domPreset from '../../../modules/presets/react-dom'
-
-
-describe('Resolving react-dom styles', () => {
-})
