--- test/mocha.opts
@@ -1,2 +1,0 @@
---compilers js:test/support/compiler.js
---recursive
